backend/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.9
channels==4.1.0
channels-redis==4.2.0
python-dotenv==1.0.1
